changed productive hours and cross utilized hours to decimal

+++ OMT.DataAccess/Entities/Prod_Util.cs
@@ -13,8 +13,8 @@ namespace OMT.DataAccess.Entities
         public int Prod_UtilId { get; set; }
         public int AgentUserId{ get; set; }
         public int Total_Shift_Hours { get; set; }
-        public int Productive_Hours { get; set; }
-        public int Cross_Utilized_Hours { get; set; }
+        public decimal Productive_Hours { get; set; }
+        public decimal Cross_Utilized_Hours { get; set; }
         public int Productivity_Percentage { get; set; }
         public int Utilization_Percentage { get; set; }
         public int CrossUtilization { get; set; }
